requirements.txt
pytelegrambotapi==4.7.1
schedule~=1.1.0
selenium~=4.5.0
Flask~=2.2.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
SQLAlchemy==1.4.39
PyJWT~=2.6.0
Werkzeug~=2.2.1
flask-restx==1.0.6
undetected-chromedriver==3.4.6
python-dotenv==0.21.1